+++ Common/EntityEffects/NpcBodyRotation.cs
@@ -11,7 +11,7 @@ namespace TerrariaOverhaul.Common.EntityEffects;
 
 public sealed class NpcBodyRotation : GlobalNPC
 {
-	public static readonly ConfigEntry<bool> EnableEnemyTiltingEffects = new(ConfigSide.ClientOnly, true, "Visuals");
+	public static readonly ConfigEntry<bool> EnableEnemyTiltingEffects = new(ConfigSide.ClientOnly, true, "Visuals", "Enemies");
 
 	private float usedRotationOffset;