+++ Common/RngRemoval/RemoveDamageRandomization.cs
@@ -1,11 +1,14 @@
 ﻿using System;
 using Terraria;
 using Terraria.ModLoader;
+using TerrariaOverhaul.Core.Configuration;
 
 namespace TerrariaOverhaul.Common.RngRemoval;
 
 public sealed class RemoveDamageRandomization : ILoadable
 {
+	public static readonly ConfigEntry<bool> EnableDamageRandomizationRemoval = new(ConfigSide.Both, true, "Balance");
+
 	public void Load(Mod mod)
 	{
 		On_Main.DamageVar_float_int_float += DamageVarDetour;
@@ -14,5 +17,11 @@ public sealed class RemoveDamageRandomization : ILoadable
 	public void Unload() { }
 
 	private static int DamageVarDetour(On_Main.orig_DamageVar_float_int_float orig, float damage, int percent, float luck)
-		=> (int)Math.Round(damage);
+	{
+		if (!EnableDamageRandomizationRemoval) {
+			return orig(damage, percent, luck);
+		}
+
+		return (int)Math.Round(damage);
+	}
 }